Add instructions for building on CSIL machine
Add the `csil-build` directory, containing instructions and simple
scripts for building CMPT 756 service images on a CSIL machine.

These are most useful for users with ARM machines (such as the Apple
M1), who cannot build the images on their own machines.

The build process does not run inside the CMPT 756 tools container but
instead runs directly on the CSIL OS, which as of Nov 2021 is Ubuntu
20:04. The tool chain (Ubuntu, make, docker) is sufficiently similar
to ours, as well as standardized across all CSIl machines, there
should be no difference from running on the tools container.

We cannot run the tools container because the CSIL security regime
does not permit running `docker` inside a container.