Map ~/.minikube volume to tools container
This is a necessary but not sufficient step to access Minikube from
the tools container. In addition, we likely have to also map this
directory into $(HOME)/.minikube in the *container*'s filespace, where
$(HOME) is the path to the user's home directory in the *host*
filespace. This is because some files in .minikube have absolute
host paths to the other files (including the .pem keys) and so the host
paths must also exist in the container filespace.

This still isn't enough.  For reasons not yet diagnosed, even when
`kubectl` can find all required Minikube files, the Minikube control
plane will not recognize calls from inside the tools container.  This
may be due to Minikube settings on what IP addresses can access the
control plane or due to limitations of networking from inside a
container out to wherever the control plane is running (a VM or
another container).