Reorder options to sed
"sed" seems to be especially wide variations in how different versions
handle their options.  The version included in the tools container
appears to require that '-i' be the *last* option on the line. If
instead we have a sequence such as '-i -e ....' (the order before
this commit), the '-e' is mis-interpreted as a requst to backup the
original file with that suffix.

This commit moves '-i' to just before the file name, preventing that
mis-interpretation.