Allow override of default server IP address
The default server IP address is the address in the bridge network of
the container named `s2`.  The make fails if no such container exists.

This commit allows the caller to override the server address using
`make SERVER=...`.  This is essential when the client is contacting a
remote server on EC2 or Kubernetes.