Remove ZZ-AWS-SESSION-TOKEN for "starter" AWS accounts
The Spring 2021 version of the course used AWS "starter"
accounts. These proved to be awkward, in particular requiring the use
and constant refreshing of session ids.  For Spring 2022, we are
requiring students to use regular, pay-for accounts.

We eliminate the ZZ-AWS-SESSION-TOKEN from the list or template
variables and modify `tools/call-sed.sh` to expect one less token. The
script also no longer has to optionally delete ZZ-AWS-SESSION-TOKEN
line whenever it is empty.