Add CI hooks to music service
Exted the music service app and music cli as well as add
a special test program for use in CI.

The server now has a "shutdown" route and a Makefile entry that
runs it in detached mode.

The cli now has a "shutdown" command that calls the associated route
in the server.

The test program simply calls the server's test route, followed
by the shutdown route. If both return 200, the test is considered
successful (return code 0), otherwise unsuccessful (return code 10.