Define the role of `aws-rotate-cred.sh`
Working further with this script, I realized that its only useful
role (and it is a useful role) is for rotating in a new key after
an old key has been deleted.

You can't use it for a brand-new key because the script can
only be run when the user already has an active key.

Note that inactivating and deleting the old key still has to be
done by the user (either via the console or command-line) before
running this script.