Remove share of ~/.git with the tools container
The tools/shell.sh call shared ~/.git with the tools container.
This was intended to share the Git configuration but that is in fact
in ~/.gitconfig.  Also, sharing that configuration may break Git
because configuration entries that are appropriate in the host OS
may not work in the guest container (due to paths that only
exist in the host OS, Git extensions only installed in the host OS,
and similar items).

Instead, users of the tool container should develop their own
way of setting Git configuration in the tools container or simply
not use Git in the container, only in the host OS.