Add A3 code
Assignment 3 is based on the common s2/standalone infrastructure.  It
adds files and updates as necessary to pull the music service down to
the remote instance and run it.

The commit adopts the convention that each of A1--A3 will have
different source files, suffixed by their assignment (`-a3.py`).
Assignments 1 and 2 will have this change made in future commits.

This commit updates builda3.sh and Makefile to build the A3 version
similarly to A2 but students are requested not to use these.  Instead,
they should use docker commands directly, as specified in the
written assignment.

The commit also includes some code and infrastructure for building
different versions of the image, including one or two extra layers with
a file of 4KiB random bytes.  These are not currently used in the
written assignment but retained here in case they prove useful
later.