Use distinct source file names for A1--A3
The original exercises reused the name `app.py` for Assignments 1, 2,
and 3.  This made it difficult to see what a student had done for
Assignment 1 once they had pushed commits for Assignments 2 or 3.

This commit renames the source files app-a1.py, app-a2.py, app-a3.py.
The Dockerfile renames each to app.py when building the image. This
requires a new Dockerfile argument, ASSIGN, which in turn requires
a new Makefile argument, ASSIGN.

This commit has been tested as completely as possible but there may
be lingering cases where the old name, app.py, remains in use.