Separate out A1-A3 CI status badges
Move the links to the A1--A3 status badges into a templated README.md
in the s2/standalone directory.  The template is required because
GitHub (apparently) requires a full URL path to the badges, including
the owner of the current repo.  When instantiating the course template,
the original owner, scp756-221, is copied straight over in any
untemplated file.

Templating the top-level README.md makes editing it awkward and clutters
the top level of the repo.  The status codes are specific to
s2/standalone in any case, so it makes sense to push them there.
Isolated in the s2/standalone README, there is no editing interference
templating the file.