Update allclouds-tpl.mak to handle missing CLIs
The allclouds-tpl.mak makefile calls the vendor CLI tools
eksctl, az, and gcloud.  We are guaranteed to have eksctl
installed by the tools container, but the other two may be
missing.

This commit modifies the makefile so that it checks for the
existence of az and gcloud and only calls them if they exist.
If one does not exists, a message is printed in lieu of calling
it.