Add CI framework
The ci subdirectory contains the basic framework of a CI system. It
uses docker-compose to set up dynamodb-local, db, s1, s2, and a
test system.

db/app-tpl.py had to be modified to check if the environment
includes a `DYNAMODB_URL` environment variable. If found,
the DB service uses that URL (which for CI will point to the
local DynamoDB rather than the AWS service).

The code requires the update to the template tool that sets the
executable flag of the output file if the source file is
executable.