Remove lingering reference to ZZ-AWS-SESSION-TOKEN
The process-templates.sh script retained code that dealt with
the possibility that ZZ-AWS-SESSION-TOKEN was not set. It
passed an empty argument to call-sed.sh to ensure that a value
was always there.

Now that the session token is deleted, we no longer need that
mechanism.

This also means that call-sed.sh can require an exact number of
arguments (currently 6), rather than a minimum.