Make output file executable if template source is
Add a step to call-sed.sh that makes the output file executable if
the source template is.

The required mechanism is awkward---`test -x` does not seem to work
correctly on shared files if the host OS is macOS.  Instead of `test`,
the text output of `ls -l` is examined.  This is slow but it works
even for guest Linux containers.