Return artist and title from music.read()
The old music.read() simply returned the status code but not
the values of the record actually read.

This commit extends music.read() to return the artist and title.
It extends ci_test to compare the returned artist and title with
what was passed to music.create() when the song was added.