Add trailing slash to . for multi-file COPY
The version of `docker` used in GitHub is more stringent in its
syntax checking than the version used in development. Apparently
the latest requirement for Dockerfile is that any COPY command
with more than one source file must specify the destination directory
with a trailing slash. This applies even when the destination is
the current directory.  Thus the target '.' must be written as './'.

Side note:  Visual Studio Code had underlined these usages, indicating
incorrect syntax, but I ignored that because I was confident that
a simple '.' was obviously correct.

The more you know!