Add directory to extend the tools with gcloud
This directory contains the tools for extending the tools container image
with the `gcloud` command-line interface.

See the `README.md` for how to use these tools.

NOTE: The Dockerfile in this commit pulls the `v1.0beta4` version
of the tools image.  It should eventually be updated to pull the latest
version used in the actual course.