Install docker-compose v2 in system-side CI
Apparently (see the comments in `ci-system-v1.1.yaml`) we have to
install docker-compose v2 if we are to guarantee its presence at
a known path.

Whatever, man.  I am so over this docker-compose version drama.

This commit adds the install step to the v1.1 CI action. It
also modifies runci.sh to accept an environment variable that
specifies the path to the downloaded.  This variable is only
set in GH Actions, to ensure we get the version we oh-so-carefully
downloaded.