Check flake8 conformance in CI tests
This commit adds flake8 conformance checks to the `ci-system-v*.yaml`
tests.  This requires the following changes:

1. Install the appropriate version of flake8 in the GitHub runners
   (see .github/workflows/ci-system-v*.yaml`)
2. Invoke flake8 from `runci.sh`.
3. Specify the relevant dirs to check via a `flake-dirs.txt` for
   each CI version subdirectory.

Add the infrastructure for running flake8 on all Python code
when doing CI.

These tests require that the GitHub runner have flake8 4.0.1 installed,
which is (probably?) not the default.

This commit installs that version of flake8 for the v1 test but not the
v1.1.  This is to see if the v1.1 will fail in the absene of a flake8
install.