Add "c756 quickies" to profiles
This commit adds the AWS features of the "756 quickies" to the
`profiles` directory. This adds default bash and Git aliases. It
also adds the AWS aliases and EC2 "package" system, required
in the GPU assignment.

Students still have to do a modicum of configuration before using
the Git aliases (defining their name and email) and the AWS aliaes
(EC2 key pair and security group).